+++ src/Redux/Car/carSlice.js
@@ -20,6 +20,22 @@ export const getCars = createAsyncThunk(
   },
 );
 
+export const getCarDetail = createAsyncThunk(
+  'car/getCarDetail',
+  (id) => {
+    const result = fetch(
+      `http://localhost:3000/api/v1/car/${id}`,
+      {
+        method: 'get',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      },
+    ).then((response) => response.json());
+    return result;
+  },
+);
+
 export const carSlice = createSlice({
   name: 'cars',
   initialState,
@@ -29,6 +45,11 @@ export const carSlice = createSlice({
     [getCars.fulfilled]: (state, action) => ({
       ...state,
       cars: action.payload.cars,
+    }),
+
+    [getCarDetail.fulfilled]: (state, action) => ({
+      ...state,
+      car: action.payload.car,
     }),
   },
 });
